docs: Add a reference to the data type conversion
Since version 7.9 it is possible to cast a conversion from string to one
of the supported data types (integer, long, float, double, boolean or
ip). Since we upgraded to the latest version in #15 we now added a small
paragraph text below the form that points users to the new feature.