Show the backend version of the instance (#16)
* beats: Show the backend version of the instance

We're now showing the beats version used to generate the current build.
This is both shown in the startup logs and in the Web UI. This should
inform users of the features available in the current instance. In
practice the dissect filter doesn't change that often.

* Avoid initializing a new struct for each request

Since the beats version information is constant and defined at build
time, we create a global `versionInfo` struct for the lifecycle of the
entire app.